+++ GROD/GROD.csproj
@@ -9,17 +9,17 @@
     <Company>BakkerGames</Company>
     <Copyright>Copyright 2023-2024 Scott Bakker</Copyright>
     <Description>GROD - Game Resource Overlay Dictionary
-An in-memory key-value dictionary for handling starting data and subsequent changes. Designed for games needing save/restore ability.</Description>
+An in-memory key-value dictionary for handling beginning data and subsequent changes. Designed for games needing save/restore ability.</Description>
     <PackageLicenseExpression>MIT-0</PackageLicenseExpression>
     <PackageOutputPath>..\NuGet\</PackageOutputPath>
-    <PackageProjectUrl>https://github.com/BakkerGames/GrodLibrary</PackageProjectUrl>
+    <PackageProjectUrl>https://github.com/BakkerGames/GROD</PackageProjectUrl>
     <PackageReadmeFile>README.md</PackageReadmeFile>
-    <RepositoryUrl>https://github.com/BakkerGames/GrodLibrary</RepositoryUrl>
+    <RepositoryUrl>https://github.com/BakkerGames/GROD</RepositoryUrl>
     <Title>GROD - Game Resource Overlay Dictionary</Title>
-    <Version>1.0.6</Version>
-    <AssemblyVersion>1.0.6</AssemblyVersion>
-    <FileVersion>1.0.6</FileVersion>
-    <PackageTags>grod;key value;data;overlay;game;game resources;gamedev;game development;interactive fiction;IF;text adventure</PackageTags>
+    <Version>1.1.0</Version>
+    <AssemblyVersion>1.1.0</AssemblyVersion>
+    <FileVersion>1.1.0</FileVersion>
+    <PackageTags>grod;data;overlay;game;resources;gamedev;development</PackageTags>
     <PackageReleaseNotes></PackageReleaseNotes>
   </PropertyGroup>
